rename eqsn, fixed spelling mistake
